booking/requirement.txt
decorator==4.4.2
Django==2.2
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.0
PyMySQL==0.10.1
pytz==2020.1
sqlparse==0.4.1
